etna/packages/etna-py/pyproject.toml
[tool.poetry]
name = "mountetna"
version = "0.1.1"
description = "Client package for Mount Etna data library"

license = "GPL-2.0-only"

authors = [
    "Zachary Collins <zachary.collins@ucsf.edu>",
    "Saurabh Asthana <saurabh.asthana@ucsf.edu>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/mountetna/monoetna/"
homepage = "https://github.com/mountetna/monoetna"

keywords = ['mount etna', 'data library']

[tool.poetry.dependencies]
python = "^3.8"  # Compatible python versions must be declared here
pandas = "^1.2.1"
requests = "^2.28.0"
cryptography = "^37.0.3"
pyserde = "^0.8.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
